requirements.txt
albumentations
scikit_image
opencv_python
opencv-python-headless
tqdm
numpy
matplotlib
Pillow
PyYAML
# scikit-learn
tensorboard
mlconfig